requirements.txt
Pillow>=7.1.2
pkg-resources==0.0.0
